+++ apps/i15-1/src/main.tsx
@@ -1,26 +1,14 @@
-import { DiamondTheme, ThemeProvider } from "@diamondlightsource/sci-react-ui";
+import { DiamondTheme } from "@diamondlightsource/sci-react-ui";
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
 import { createRoot } from "react-dom/client";
-import { StrictMode, useEffect, useState } from "react";
+import { StrictMode } from "react";
 
 import { Layout } from "./routes/Layout.tsx";
 import Dashboard from "./routes/Dashboard.tsx";
 import Robot from "./routes/Robot.tsx";
-import { InstrumentSessionProvider } from "./context/instrumentSession/InstrumentSessionProvider.tsx";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { store } from "@diamondlightsource/cs-web-lib";
-import { Provider } from "react-redux";
-
-declare global {
-  interface Window {
-    global?: typeof globalThis;
-  }
-}
 
 import { createApi } from "@atlas/blueapi";
-import { BlueapiProvider } from "@atlas/blueapi-query";
-import { useLoadPvwsConfig } from "@atlas/pvws-config";
-import { UserAuthProvider } from "./context/userAuth/UserAuthProvider.tsx";
+import { AppProviders } from "./AppProviders.tsx";
 
 async function enableMocking() {
   if (import.meta.env.DEV) {
@@ -47,32 +35,13 @@ const router = createBrowserRouter([
 ]);
 
 const api = createApi("/api");
-const queryClient = new QueryClient();
-
-function App() {
-  const config = useLoadPvwsConfig();
-
-  return (
-    <Provider store={store(config)}>
-      <QueryClientProvider client={queryClient}>
-        <UserAuthProvider>
-          <BlueapiProvider api={api}>
-            <RouterProvider router={router} />
-          </BlueapiProvider>
-        </UserAuthProvider>
-      </QueryClientProvider>
-    </Provider>
-  );
-}
 
 enableMocking().then(() => {
   createRoot(document.getElementById("root")!).render(
-    <InstrumentSessionProvider>
-      <StrictMode>
-        <ThemeProvider theme={DiamondTheme} defaultMode="light">
-          <App />
-        </ThemeProvider>
-      </StrictMode>
-    </InstrumentSessionProvider>,
+    <StrictMode>
+      <AppProviders api={api} theme={DiamondTheme}>
+        <RouterProvider router={router} />
+      </AppProviders>
+    </StrictMode>,
   );
 });
